feat: hoc авторизации
