+++ packages/client/src/App.tsx
@@ -1,29 +1,9 @@
-import { BackgroundLayout } from './layouts/BackgroundLayout';
-import { useAppDispatch, useAppSelector } from './hooks/redux';
 
-import { fetchAuth } from './store/auth/AuthActions';
-import { Spinner } from './components/Spinner';
 import { ErrorsNotification } from './components/ErrorsNotification';
 import { Routing } from './components/Routing';
-import { useMountEffect } from './hooks/useMountEffect';
-import { selectCheckAuth } from './store/auth/AuthSelectors';
+import { withAuth } from './hocs/withAuth'; 
 
-export const App = () => {
-  const dispath = useAppDispatch();
-  const checkAuth = useAppSelector(selectCheckAuth);
-
-  useMountEffect(() => {
-    dispath(fetchAuth());
-  });
-
-  if (!checkAuth) {
-    return (
-      <BackgroundLayout>
-        <Spinner />
-      </BackgroundLayout>
-    );
-  }
-
+ const AppComponent = () => {
   return (
     <>
       <ErrorsNotification />
@@ -31,3 +11,6 @@ export const App = () => {
     </>
   );
 };
+
+
+export const App = withAuth(AppComponent);
