+++ packages/client/src/store/auth/AuthActions.ts
@@ -4,6 +4,7 @@ import { SignupData, SigninData } from 'src/modules/IAuth';
 
 import { fetchApi } from '../utils';
 import { addError } from '../error/ErrorSlice';
+import { IUser } from 'src/modules/IUser';
 
 const defaultHeaders = {
   'content-type': 'application/json',
@@ -13,8 +14,8 @@ const defaultHeaders = {
 export const fetchAuth = createAsyncThunk(
   'auth/fetchAuth',
   async (_, thunkApi) => {
-    try {
-      return await fetchApi('/auth/user', {});
+    try { 
+      return await fetchApi<IUser>('/auth/user', {});  
     } catch (error) {
       return thunkApi.rejectWithValue('Ошибка авторизации');
     }
